Adds unique file name function. Changes return type of the bmp function (returns filesize as size_t instead of int). Tries to fix the red-blue-flip-bug with SKönigs approach (didn't worked so far)

+++ bmp.h
@@ -18,6 +18,6 @@
 
 typedef unsigned char uchar;
 
-int bmpWriteColor(uchar *buffer, int width, int height, char *filename);
+size_t bmpWriteColor(uchar *buffer, int width, int height, char *filename);
 
 #endif
